pyproject.toml
[project]
name = "binding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "einops>=0.8.1",
    "glasbey>=0.3.0",
    "hydra-core>=1.3.2",
    "lmdeploy>=0.10.2",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "omegaconf>=2.3.0",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pyrootutils>=1.0.4",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "sentencepiece>=0.2.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
]
